requirements.txt
foolbox==3.3.4
matplotlib==3.9.1
numpy==2.0.0
pillow==10.4.0
scikit-image==0.24.0
scikit-learn==1.5.2
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.66.4
wandb==0.17.4
diffusers==0.29.2
opencv-python==4.10.0.84